Export k8s flags
