+++ pkg/kubeflags/kubeflags.go
@@ -1,0 +1,10 @@
+package kubeflags
+
+type EnableFlag string
+
+const (
+	EnabledTrue       = EnableFlag("true")
+	EnabledFalse      = EnableFlag("false")
+	EnabledAutodetect = EnableFlag("autodetect")
+	EnabledDefault    = EnabledFalse
+)
